Added tests for Jump And Link instructions, fixed alignment for jalr according to specification.
