+++ src/main/java/com/cburch/logisim/riscv/rv32imData.java
@@ -164,12 +164,12 @@ class rv32imData implements InstanceData, Cloneable {
         break;
       case 0x6F:  // Jump And Link (J-Type)
         JumpAndLink.link(this); //jal rd,label
-        pc.set((pc.get() + ir.imm_J()) & 0xffffffff);
+        pc.set((pc.get() + ir.imm_J()) & 0xffffffffL);
         fetchNextInstruction();
         break;
       case 0x67:  // Jump And Link Reg (I-Type)
         JumpAndLink.link(this); //jalr rd,rs1,imm_I
-        pc.set((getX(ir.rs1()) + ir.imm_I()) & 0xffffffff);
+        pc.set((getX(ir.rs1()) + ir.imm_I()) & 0xfffffffeL);
         fetchNextInstruction();
         break;
       case 0x37:  // Load Upper Immediate (U-type)
